==> text classification/00_eda.ipynb <==
### This notebook is basically scratch paper.  We were just playing around with the data and seeing what is there.  Additionally, we tested different methods and functions to see what they would do on this data.  Everything that we used to clean and process the data was copied to 01_datacleaning.

import pandas as pd
import numpy as np
from sklearn.feature_extraction.text import CountVectorizer
from nltk.stem import WordNetLemmatizer, SnowballStemmer
from nltk.stem.porter import PorterStemmer
import re

# https://songhuiming.github.io/pages/2017/04/02/jupyter-and-pandas-display/
pd.options.display.max_rows
pd.set_option('display.max_colwidth', -1)

train = pd.read_csv('./data/train.csv')
train.head()

train.isna().sum()

train.shape

train.dtypes

train[train['location'].notna()].head()

train['keyword'].value_counts()

train[train['keyword'].isna()]['target'].value_counts()

train['text'].str.contains('forest').sum()

train[(train['keyword'].notna()) & (train['text'].str.contains('forest fire'))]

train[(train['text'].str.contains('#'))]

train['text'] = train['text'].str.lower()
train['keyword'] = train['keyword'].str.lower()

len('6ioabsl6i7')

test = pd.read_csv('./data/test.csv')
test.head()

train[train['text'].duplicated()]

test['text'].duplicated().sum()

train[(train['text'].str.contains('t.co'))]['target'].value_counts(normalize=True)

train['target'].value_counts(normalize=True)

train[(train['text'].str.contains('http'))&~(train['text'].str.contains('t.co'))]

train[train['keyword'].str.contains(' ')]

keyword_dis = {}
for keyword in set(train['keyword']):
    keyword_dis[keyword] = train[train['keyword'] == keyword]['target'].mean()

for key, value in keyword_dis.items():
    if value > .9:
        print(f'{key} : {value:.4}')

for key, value in keyword_dis.items():
    if value < .25:
        print(f'{key} : {value:.4}')

train['keyword'].fillna('-', inplace=True)

i = train['text'].str.find('t.co')

train.iloc[68]['text'].slice[6,9]

for i in train['text'][:100]:
    if 'https' in i:
        locat = i.find('https')
        print(i[locat:locat+23])       
    elif 'http' in i:
        locat = i.find('http')
        print(i[locat:locat+22])      

URLless_string = re.sub(r'\w+:\/{2}[\d\w-]+(\.[\d\w-]+)*(?:(?:\/[^\s/]*))*', '', i)
print(URLless_string)

# Using this string remover from StackOverflow: 
# https://stackoverflow.com/questions/11331982/how-to-remove-any-url-within-a-string-in-python
train['text_nourl'] = [re.sub(r'\w+:\/{2}[\d\w-]+(\.[\d\w-]+)*(?:(?:\/[^\s/]*))*', '', i) for i in train['text']]

train[train['text'].str.contains('http')].head()

train['text_nourl'].duplicated().sum()

train['keyword'] = train['keyword'].str.replace('%20', ' ')

keywords_sorted = sorted(set(train['keyword']))
keywords_sorted

lemmatizer = WordNetLemmatizer()
keywords_lem = []
for word in keywords_sorted:
    keywords_lem.append(lemmatizer.lemmatize(word))
keywords_lem

lemmatizer.lemmatize('bloods')

comp = list(zip(keywords_lem, keywords_sorted))
for i in comp:
    if i[0] == i[1]:
        print(i)

pstem = PorterStemmer()
keywords_stem = [pstem.stem(i) for i in keywords_sorted]
comp2 = list(zip(keywords_stem, keywords_sorted))
for i in comp2:
    if i[0] != i[1]:
        print(i)

snowstem = SnowballStemmer('english')

train['keywords_stemmed'] = train['keyword'].map(lambda x: snowstem.stem(x))
train['keywords_stemmed'].value_counts()

keyword_dis1 = {}
for keyword in set(train['keywords_stemmed']):
    keyword_dis[keyword] = train[train['keywords_stemmed'] == keyword]['target'].mean()
for key, value in keyword_dis.items():
    if value > .8:
        print(f'{key} : {value:.4}')


no_key = train.loc[train['keyword'] == '-']
for index, row in no_key.iterrows():
    for word in keywords_sorted:
        if word in row['text']:
            train.loc[index, 'keyword'] = word
train[train['keyword'] == '-']
