==> text classification/03_modeling.ipynb <==
import pandas as pd
import numpy as np
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn.pipeline import Pipeline
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.svm import SVC
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.naive_bayes import MultinomialNB
from sklearn.neural_network import MLPClassifier

train = pd.read_csv('./data/train_clean.csv')
train.head()


y = train['target']
X = train[['text_nourl', 'keywords_stemmed']]
X_train, X_test, y_train, y_test = train_test_split(X, y, 
                                                    random_state=42, 
                                                    test_size=.25, 
                                                    stratify=y)

# inspiration from https://stackoverflow.com/questions/38555650/try-multiple-estimator-in-one-grid-search
# Trying both Count and Tfidf Vectorizers
# Trying 7 different models
# Have nest loops to run all 14 combinations and 
# build a DataFrame of the results including the model
# Just used default parameters here to dial in some models then will
# Gridsearch parameters next

vects = {
    'cvec' : CountVectorizer(),
    'tfidf' : TfidfVectorizer()
}
models = {
    'mnb' : MultinomialNB(),
    'knn' : KNeighborsClassifier(),
    'logreg' : LogisticRegression(),
    'dectree' : DecisionTreeClassifier(),
    'svc' : SVC(),
    'randomforest' : RandomForestClassifier(),
    'adaboost' : AdaBoostClassifier()
    
}

modeling = []
for vec in vects:
    for mode in models:
        mod = Pipeline([(vec,vects[vec])
                                   ,(mode, models[mode])])
        curmod = mod.fit(X_train['keywords_stemmed'], y_train)
        out = {
            'name' : f'{vec} + {mode}',
            'model' : curmod,
            'train_score' : curmod.score(X_train['keywords_stemmed'],
                                        y_train),
            'test_score' : curmod.score(X_test['keywords_stemmed'],
                                       y_test)
        }
        modeling.append(out)
all_mods_keywords = pd.DataFrame(modeling)

# Built the above into a function to reuse.  
# This train, test, splits the data the same every time
# This makes it easier to pass in fewer parameters

def modeling(X, y):
    X_train, X_test, y_train, y_test = train_test_split(X, y, 
                                                    random_state=42, 
                                                    test_size=.25, 
                                                    stratify=y)
    vects = {
    'cvec' : CountVectorizer(),
    'tfidf' : TfidfVectorizer()
}
    models = {
    'mnb' : MultinomialNB(),
    'knn' : KNeighborsClassifier(),
    'logreg' : LogisticRegression(),
    'dectree' : DecisionTreeClassifier(),
    'svc' : SVC(),
    'randomforest' : RandomForestClassifier(),
    'adaboost' : AdaBoostClassifier(),
    'mlpclass' : MLPClassifier()
    
}

    outp = []
    for vec in vects:
        for mode in models:
            mod = Pipeline([(vec,vects[vec])
                                       ,(mode, models[mode])])
            curmod = mod.fit(X_train, y_train)
            out = {
                'name' : f'{vec} + {mode}',
                'model' : curmod,
                'train_score' : curmod.score(X_train,
                                            y_train),
                'test_score' : curmod.score(X_test,
                                           y_test)
            }
            outp.append(out)
    return pd.DataFrame(outp)

#Results of using the text from the tweet
models_text = modeling(X['text_nourl'], y)
models_text

# Calculating the difference between test and train scores
models_text['delta'] = models_text['train_score'] - models_text['test_score']
models_text.sort_values('delta')

# Setting up gridsearch parameters for the vectorizer and the models

vec_params = [
    {'vec__ngram_range' : [(1,1), (1,2)]},
    {'vec__stop_words' : [None, 'english']}
]
lr_params = [
    {'logreg__penalty' : ['l1', 'l2', 'none']},
    {'logreg__C' : np.linspace(0.01, 1, 5)}    
]
knn_params = [
    {'knn__n_neighbors' : [5, 9, 15]},
    {'knn__p' : [1, 2]}
]
ada_params = [
    {'adaboost__n_estimators' : [10, 50, 100]}
]
rf_params = [
    {'rf__n_estimators' : [10, 50, 100, 200]},
    {'rf__max_depth' : [None, 5, 25, 50]},
    {'rf__max_features' : ['auto', 20, 50, 80]}
]
LogisticRegression

# This intentional misspelling of model function loops the 
# Vectorizers and Models again, but this time they get gridsearched
# according to the parameters above

def modl(X, y):
    X_train, X_test, y_train, y_test = train_test_split(X, y, 
                                                    random_state=42, 
                                                    test_size=.25, 
                                                    stratify=y)
    vects = {
    'cvec' : ['vec', CountVectorizer()],
    'tfidf' : ['vec', TfidfVectorizer()]
    }
    models = {
    'knn' : KNeighborsClassifier(),
    'logreg' : LogisticRegression(),
    'rf' : RandomForestClassifier(),
    'adaboost' : AdaBoostClassifier(),
       
    }
    param_dict = {
        'knn' : knn_params,
        'logreg' : lr_params,
        'rf' : rf_params,
        'adaboost' : ada_params
    }
    outp = []
    for vec in vects:
        for mode in models:
            mod = Pipeline([(vects[vec][0],vects[vec][1])
                                       ,(mode, models[mode])])
            gs = GridSearchCV(mod,
                             vec_params+param_dict[mode],
                             cv=5,
                             n_jobs = -1)
            
            curmod = gs.fit(X_train, y_train)
            gsbest = curmod.best_estimator_
            out = {
                'name' : f'{vec} + {mode}',
                'model' : gsbest,
                'train_score' : gsbest.score(X_train,
                                            y_train),
                'test_score' : gsbest.score(X_test,
                                           y_test)
            }
            outp.append(out)
    return pd.DataFrame(outp)

# Running it on keywords
models_keys_gs = modl(X['keywords_stemmed'],y)
models_keys_gs

models_keys_gs['delta'] = models_keys_gs['train_score'] - models_keys_gs['test_score']
models_keys_gs

# Saving what we deemed as the best model for this
best_model = models_keys_gs.loc[6,'model']

# Running it for the text of the tweets and saving it
models_text_gs = modl(X['text_nourl'], y)
models_text_gs['delta'] = models_text_gs['train_score'] - models_text_gs['test_score']
models_text_gs

# Building a table of all the scores for each model and test/train scores
scores = models_keys_gs[['name', 'train_score', 'test_score']].copy()
scores.rename(columns={'train_score' : 'keywords_train', 'test_score' : 'keywords_test'}, inplace=True)
scores['text_train'] = models_text_gs['train_score']
scores['text_test'] = models_text_gs['test_score']
scores

best_model_tx = models_text_gs.loc[6,'model'] #Saving best model
preds = best_model_tx.predict(X_test['text_nourl'])

# Trying to dial in further on what we deemed the best model again
pipe_tf = Pipeline([
    {'tfvec' : TfidfVectorizer()},
    {'rf' : RandomForestClassifier()}
])
pipe_params = [
    {'vec__ngram_range' : [(1,1), (1,2)]},
    {'vec__stop_words' : [None, 'english']},
    {'rf__n_estimators' : [10, 50, 100, 200]},
    {'rf__max_depth' : [None, 5, 25, 50]},
    {'rf__max_features' : ['auto', 20, 50, 80]}
    ]
gs_rf = GridSearchCV(pipe_tf,
                    pipe_params,
                    cv=7,
                    n_jobs=-1)
gs_rf.fit(X_train['text_nourl'], y_train)
print(gs_rf.best_score_)
gs_rf_best = gs_rf.best_estimator_

# Kaggle submission creation
test = pd.read_csv('./data/test_clean.csv')
preds = best_model_tx.predict(test['text_nourl'])
pred_1 = pd.DataFrame()
pred_1['id'] = test['id']
pred_1['target'] = pred
pred_1.to_csv('./data/submissions/rftext.csv', index=False)

# Kaggle submission creation
def kagg_sub (model, Xtest, name):
    preds = model.predict(Xtest)
    outp = pd.DataFrame()
    outp['id'] = test['id']
    outp['target'] = preds
    outp.to_csv(f'./data/submissions/{name}.csv', index=False)


kagg_sub(best_model, test['keywords_stemmed'], 'keywords1')
